CUF last version
bugs fixed, functionality added and fully tested!

+++ satellite-subsystems/demo/src/sub-systemCode/CUF/CUF.h
@@ -65,7 +65,7 @@ void* CUFSwitch(int index);
  * @return the generated ssh
  * 0 on fail
  */
-unsigned long GenerateSSH(char* data, int len);
+unsigned int GenerateSSH(char* data, int len);
 
 /*!
  * Function to authenticate a CUF
